fix unable to delete if image is null

+++ src/pages/Blog.jsx
@@ -3,7 +3,7 @@ import { deleteObject, ref } from 'firebase/storage'
 import React from 'react'
 import { useState } from 'react'
 import { useEffect } from 'react'
-import { useNavigate, useParams } from 'react-router-dom'
+import { Link, useNavigate, useParams } from 'react-router-dom'
 import { auth, db, storage } from '../firebase-config'
 
 function Blog({postId}) {
@@ -18,12 +18,19 @@ function Blog({postId}) {
         try {
             const getPath = ref(storage, blogPost.imageURL)
             const objRef = ref(storage, getPath.fullPath)
-            await deleteObject(objRef)
-            await deleteDoc(doc(db, 'posts', pid));
-            navigate('/')
-        } catch (error) {
-            if(error.message === 'storage/object-not-found') return
+            if  (objRef.fullPath === '') {
                 await deleteDoc(doc(db, 'posts', pid));
+                navigate('/')
+            } else {
+                await deleteObject(objRef)
+                await deleteDoc(doc(db, 'posts', pid));
+                navigate('/')
+            }
+        } catch (error) {
+            console.log(error)  
+            if(error.message === 'storage/object-not-found') {
+                await deleteDoc(doc(db, 'posts', pid));
+            }
         }
     }
 
@@ -36,7 +43,6 @@ function Blog({postId}) {
             const response = await getDoc(postRef)
             setBlogPost(response.data())
             setAuthor(response.data().author)
-            console.log(auth.currentUser.uid);
         }
         getById()
     },[id.postId])
@@ -58,7 +64,9 @@ function Blog({postId}) {
             </div>
             {author.id === auth.currentUser.uid ? 
                 <div className="blog-post__actions">
-                    <button>Edit</button>
+                    <button>
+                        <Link to={`/blog/edit/${id.postId}`}>Edit</Link>
+                    </button>
                     <button onClick={() => deletePost(id.postId)}>Delete</button>
                 </div> : ''
             }
